Separating out config
