+++ ironfish-cli/src/utils/chainport/chainport.ts
@@ -1,44 +1,16 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { Logger, MAINNET, TESTNET } from '@ironfish/sdk'
+import { Logger } from '@ironfish/sdk'
 import { CliUx } from '@oclif/core'
 import axios from 'axios'
+import { getNetworkConfig } from './config'
 import {
   ChainportBridgeTransaction,
   ChainportNetwork,
   ChainportTransactionStatus,
   ChainportVerifiedToken,
 } from './types'
-
-const config = {
-  [TESTNET.id]: {
-    endpoint: 'https://preprod-api.chainport.io',
-    outgoingAddresses: [
-      '06102d319ab7e77b914a1bd135577f3e266fd82a3e537a02db281421ed8b3d13',
-      'db2cf6ec67addde84cc1092378ea22e7bb2eecdeecac5e43febc1cb8fb64b5e5',
-      '3bE494deb669ff8d943463bb6042eabcf0c5346cf444d569e07204487716cb85',
-    ],
-    incomingAddresses: ['06102d319ab7e77b914a1bd135577f3e266fd82a3e537a02db281421ed8b3d13'],
-  },
-  [MAINNET.id]: {
-    endpoint: 'https://api.chainport.io',
-    outgoingAddresses: [],
-    incomingAddresses: [],
-  },
-}
-
-export const getNetworkConfig = (networkId: number) => {
-  if (networkId !== TESTNET.id && networkId !== MAINNET.id) {
-    throw new Error(`Unsupported network ${networkId} for chainport`)
-  }
-
-  if (networkId === MAINNET.id) {
-    throw new Error(`Mainnet is not yet supported.`)
-  }
-
-  return config[networkId]
-}
 
 export const getChainportTransactionStatus = async (networkId: number, hash: string) => {
   const config = getNetworkConfig(networkId)
